tech(server): Extract global vars to its own package
