+++ server/src/main/java/com/chutneytesting/RestExceptionHandler.java
@@ -4,7 +4,6 @@ import com.chutneytesting.admin.domain.BackupNotFoundException;
 import com.chutneytesting.admin.domain.gitbackup.UnreachableRemoteException;
 import com.chutneytesting.campaign.domain.CampaignNotFoundException;
 import com.chutneytesting.design.domain.dataset.DataSetNotFoundException;
-import com.chutneytesting.design.domain.globalvar.GlobalVarNotFoundException;
 import com.chutneytesting.environment.domain.exception.AlreadyExistingEnvironmentException;
 import com.chutneytesting.environment.domain.exception.AlreadyExistingTargetException;
 import com.chutneytesting.environment.domain.exception.EnvironmentNotFoundException;
@@ -16,6 +15,7 @@ import com.chutneytesting.execution.domain.compiler.ScenarioConversionException;
 import com.chutneytesting.execution.domain.history.ReportNotFoundException;
 import com.chutneytesting.execution.domain.scenario.FailedExecutionAttempt;
 import com.chutneytesting.execution.domain.scenario.ScenarioNotRunningException;
+import com.chutneytesting.globalvar.domain.GlobalVarNotFoundException;
 import com.chutneytesting.scenario.domain.AlreadyExistingScenarioException;
 import com.chutneytesting.scenario.domain.ScenarioNotFoundException;
 import com.chutneytesting.scenario.domain.ScenarioNotParsableException;
