+++ server/src/main/java/com/chutneytesting/admin/infra/storage/FileSystemBackupRepository.java
@@ -10,11 +10,11 @@ import com.chutneytesting.admin.domain.BackupRepository;
 import com.chutneytesting.admin.domain.Backupable;
 import com.chutneytesting.admin.domain.HomePageRepository;
 import com.chutneytesting.agent.domain.explore.CurrentNetworkDescription;
-import com.chutneytesting.design.domain.globalvar.GlobalvarRepository;
-import com.chutneytesting.scenario.infra.compose.orient.OrientComponentDB;
 import com.chutneytesting.environment.domain.Environment;
 import com.chutneytesting.environment.domain.EnvironmentRepository;
+import com.chutneytesting.globalvar.domain.GlobalvarRepository;
 import com.chutneytesting.jira.domain.JiraRepository;
+import com.chutneytesting.scenario.infra.compose.orient.OrientComponentDB;
 import com.chutneytesting.tools.Try;
 import com.chutneytesting.tools.ZipUtils;
 import com.chutneytesting.tools.file.FileUtils;
