+++ server/src/main/java/com/chutneytesting/execution/domain/compiler/ComposedTestCasePreProcessor.java
@@ -1,9 +1,9 @@
 package com.chutneytesting.execution.domain.compiler;
 
 import com.chutneytesting.design.domain.dataset.DataSetRepository;
-import com.chutneytesting.design.domain.globalvar.GlobalvarRepository;
 import com.chutneytesting.execution.domain.ExecutionRequest;
 import com.chutneytesting.execution.domain.scenario.composed.ExecutableComposedTestCase;
+import com.chutneytesting.globalvar.domain.GlobalvarRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Component;
