+++ server/src/test/java/com/chutneytesting/admin/infra/storage/FileSystemBackupRepositoryTest.java
@@ -15,11 +15,11 @@ import com.chutneytesting.admin.domain.BackupNotFoundException;
 import com.chutneytesting.admin.domain.BackupRepository;
 import com.chutneytesting.admin.domain.HomePageRepository;
 import com.chutneytesting.agent.domain.explore.CurrentNetworkDescription;
-import com.chutneytesting.design.domain.globalvar.GlobalvarRepository;
-import com.chutneytesting.scenario.infra.compose.orient.OrientComponentDB;
 import com.chutneytesting.environment.domain.Environment;
 import com.chutneytesting.environment.domain.EnvironmentRepository;
+import com.chutneytesting.globalvar.domain.GlobalvarRepository;
 import com.chutneytesting.jira.domain.JiraRepository;
+import com.chutneytesting.scenario.infra.compose.orient.OrientComponentDB;
 import com.chutneytesting.tools.Try;
 import java.io.IOException;
 import java.nio.file.Files;
