+++ server/src/test/java/com/chutneytesting/execution/domain/compiler/ComposedTestCasePreProcessorTest.java
@@ -14,14 +14,14 @@ import static org.mockito.Mockito.when;
 import com.chutneytesting.WebConfiguration;
 import com.chutneytesting.design.domain.dataset.DataSet;
 import com.chutneytesting.design.domain.dataset.DataSetRepository;
-import com.chutneytesting.design.domain.globalvar.GlobalvarRepository;
-import com.chutneytesting.scenario.domain.TestCaseMetadataImpl;
-import com.chutneytesting.scenario.domain.compose.Strategy;
 import com.chutneytesting.execution.domain.ExecutionRequest;
 import com.chutneytesting.execution.domain.scenario.composed.ExecutableComposedScenario;
 import com.chutneytesting.execution.domain.scenario.composed.ExecutableComposedStep;
 import com.chutneytesting.execution.domain.scenario.composed.ExecutableComposedTestCase;
 import com.chutneytesting.execution.domain.scenario.composed.StepImplementation;
+import com.chutneytesting.globalvar.domain.GlobalvarRepository;
+import com.chutneytesting.scenario.domain.TestCaseMetadataImpl;
+import com.chutneytesting.scenario.domain.compose.Strategy;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.Instant;
 import java.util.HashMap;
