+++ server/src/test/java/com/chutneytesting/execution/domain/compiler/DataSetPreProcessorTest.java
@@ -4,8 +4,9 @@ import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.chutneytesting.execution.domain.ExecutionRequest;
+import com.chutneytesting.globalvar.domain.GlobalvarRepository;
 import com.chutneytesting.scenario.api.raw.mapper.GwtScenarioMapper;
-import com.chutneytesting.design.domain.globalvar.GlobalvarRepository;
 import com.chutneytesting.scenario.domain.TestCaseMetadataImpl;
 import com.chutneytesting.scenario.domain.gwt.GwtScenario;
 import com.chutneytesting.scenario.domain.gwt.GwtStep;
@@ -13,7 +14,6 @@ import com.chutneytesting.scenario.domain.gwt.GwtStepImplementation;
 import com.chutneytesting.scenario.domain.gwt.GwtTestCase;
 import com.chutneytesting.scenario.domain.gwt.Strategy;
 import com.chutneytesting.scenario.domain.raw.RawTestCase;
-import com.chutneytesting.execution.domain.ExecutionRequest;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
